+++ src/lib/CMakeLists.txt
@@ -60,7 +60,6 @@ add_subdirectory(mathlib EXCLUDE_FROM_ALL)
 add_subdirectory(mixer_module EXCLUDE_FROM_ALL)
 add_subdirectory(motion_planning EXCLUDE_FROM_ALL)
 add_subdirectory(npfg EXCLUDE_FROM_ALL)
-add_subdirectory(open_drone_id EXCLUDE_FROM_ALL)
 add_subdirectory(perf EXCLUDE_FROM_ALL)
 add_subdirectory(fw_performance_model EXCLUDE_FROM_ALL)
 add_subdirectory(pid EXCLUDE_FROM_ALL)
